modifs 12/01
premier système de filtre, à voir pour répérer noms de lieux et autres
